Favorites Added
- Now tracks visited subreddits and assembles a favorites screen with
the top 6 subreddits. The favorites screen is displayed on startup or
when nothing is entered (press enter when the input field is empty).
- Favorites and the statistics are being saved to localstorage if
allowed to and will be read and parsed when you visit later.
- An error is fixed where RC tried to access the imgur API to display
the first image of an album. This didn't work for some reason and might
get repaired in the future, but the feature is now disabled and will no
longer spam the console with errors and slow down the process.