Album Accessing, 2 of 2
- Albums can now be completely accessed! When viewing an Album in the
Slideshow, press the Up Arrow. There is also a prompt that explains
this.
- In Album View, the left and right arrow keys are used to navigate,
with similar mechanics like the blue highlighting in the overview. The
up arrow exits the Album view.
- There are now four additional prompts explaining the controls
contextually. Next page in Slideshow view, Up Arrow to access album,
Left and Right arrows to navigate in album view, up arrow to exit album
view.